pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ra"
version = "1.0.0"
description = "a8ra Research Accelerator — Detection Engine"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0",
    "duckdb>=0.9",
    "pyarrow>=14.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
